PRコメント対応: Widgetbook構成をgenerator化
