+++ lib/widgetbook/app.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+import 'package:no_zan_lane/widgetbook/app.directories.g.dart';
+import 'package:widgetbook/widgetbook.dart';
+import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
+
+/// Widgetbook のルートウィジェットです。
+@widgetbook.App()
+class WidgetbookApp extends StatelessWidget {
+  /// [WidgetbookApp] を生成します。
+  const WidgetbookApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final darkTheme = ThemeData(
+      useMaterial3: true,
+      colorScheme: ColorScheme.fromSeed(
+        seedColor: Colors.teal,
+        brightness: Brightness.dark,
+      ),
+    );
+
+    return Widgetbook.material(
+      lightTheme: darkTheme,
+      darkTheme: darkTheme,
+      themeMode: ThemeMode.dark,
+      directories: directories,
+    );
+  }
+}
