+++ lib/widgetbook/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:no_zan_lane/widgetbook/widgetbook_app.dart';
+import 'package:no_zan_lane/widgetbook/app.dart';
 
 void main() {
   runApp(const WidgetbookApp());
